Alter export file name to include "all" with to and from dates
In the case where all entry groups are exported. If all entry groups
are exported, the to and from dates will represent the earliest and
most recent entry group dates.

In the case where specific dates are exported, file name will include
"entry groups" with to and from dates being the to and from dates
chosen by the used on the command line, not necessarily the
earliest and most recent entry group dates.